Update File Reading and UI
Updated file reading for track orders and login.  This is a general
format that should be able to be adapted to most of the projects file
reading.  In addition the file format has be revamped and now allows for
multiple word entries, for example; a first and last name in the name
collumn.  The UI has been further setup now to be fleshed out as project
completion progresses.